Config Pokemon Detail PokemonInfo config part. 3.

+++ app/src/main/java/com/example/pokedex/presentation/screens/pokemon_details/PokemonInfo.kt
@@ -17,8 +17,8 @@ import com.example.pokedex.R
 
 @Composable
 fun PokemonInfo(
-    weight: Long,
-    height: Long,
+    weight: Float,
+    height: Float,
 ) {
     Card(
         modifier = Modifier
